refactor(#27): 도메인 중심 패키지 구조로 전면 재편
- 레이어드 구조(controller/service/repository/dto)를 도메인 중심 구조로 전환
- chat 도메인: domain/chat/controller, domain/chat/dto 패키지로 이동
- lobby 도메인: domain/lobby/controller, dto, repository, service 패키지로 이동
- RedisPublisher, RedisSubscriber: 도메인 무관 인프라 컴포넌트로 global/redis로 이동
- 전체 주석 처리된 GuestHandshakeInterceptor.java 삭제
- 각 파일 패키지 선언 및 import 경로 전면 수정fix(#27): RedisConfig Jackson 직렬화 설정 버그 수정 - JsonMapper 중복 import 제거
