--- src/main/java/io/github/ascrew/monomatbe/controller/LobbyEventController.java
@@ -1,32 +1,0 @@
-package io.github.ascrew.monomatbe.controller;
-
-import io.github.ascrew.monomatbe.service.LobbyEventService;
-import java.security.Principal;
-import lombok.RequiredArgsConstructor;
-import org.springframework.messaging.handler.annotation.DestinationVariable;
-import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.stereotype.Controller;
-
-@Controller
-@RequiredArgsConstructor
-public class LobbyEventController {
-
-  private final LobbyEventService lobbyEventService;
-
-  // 로비가 생성될 때 로비 리스트을 보고있는 모든 클라이언트에게 로비 리스트를 새로고침하라는 메시지를 보냄
-  // 차후 유저의 수가 증가할 경우 많은 요청이 발생할 수 있으므로
-  // 일정시간(예: 5초) 동안 1건 이상 로비 생성 이벤트가 발생할 경우에만 로비 리스트 새로고침 메시지를 보내도록 개선할 수 있음
-  // 클라이언트 송신 경로 : /app/lobby/create
-  @MessageMapping("/lobby/create")
-  public void notifyLobbyListRefresh(Principal principal) {
-    lobbyEventService.notifyLobbyListRefresh();
-  }
-
-  // 로비 내부 정보가 변경될 때 해당 로비에 참여한 클라이언트들에게 로비 정보를 새로고침하라는 메시지를 보냄
-  // 로비 내부 정보 변경의 기준: 유저 입장, 유저 퇴장, 유저 준비, 유저 준비 해제 등
-  // 클라이언트 송신 경로 : /app/lobby/{code}/update
-  @MessageMapping("/lobby/{code}/update")
-  public void notifyLobbyInfoRefresh(@DestinationVariable String code, Principal principal) {
-    lobbyEventService.notifyLobbyInfoRefresh(code, principal);
-  }
-}
