--- src/main/java/io/github/ascrew/monomatbe/domain/packageInfo.java
@@ -1,9 +1,0 @@
-package io.github.ascrew.monomatbe.domain;
-
-/*
-  멀티플레이어 게임 진행을 위한 '로비(Lobby)' 데이터를 담당하는 JPA 엔티티(Entity) 클래스입니다.
-
-  데이터베이스 테이블과 1:1로 매핑되는 핵심 비즈니스 모델입니다.
-  단순한 데이터 구조체를 넘어, 입장 인원 제한, 게임 시작 상태 변경 등
-  객체 지향적으로 상태와 행위를 캡슐화하여 도메인의 핵심 비즈니스 규칙을 수호합니다.
- */
